% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptPrevalence.R
\name{calculate}
\alias{calculate}
\title{Calculates the counts of standard and source concepts}
\usage{
calculate(connectionDetails, cdmName, cdmDatabaseSchema,
  vocabDatabaseSchema, resultDatabaseSchema)
}
\arguments{
\item{cdmName}{The name of the CDM}

\item{cdmDatabaseSchema}{The schema holding the CDM data}

\item{vocabDatabaseSchema}{The schema holding the OMOP Vocabularies}

\item{resultDatabaseSchema}{The schema with the writing access}

\item{connectioDetails}{The connections details for connecting to the CDM}

\item{target_database_schema}{The schema with the writing access}
}
\description{
Calculates the counts of standard and source concepts
}
